refactor(governance): split ClickHouse repository by concern, named params, harden edge cases
Splits the 787-line ActivityMonitorClickHouseRepository into
spend/events/health repositories plus a shared schemas module, each under
the repo's 300-line SRP limit, with a thin facade preserving the existing
call surface. All 13 methods move from positional (ch, params) to a single
named-object param per project convention.

Also hardens two edge cases flagged in review: findSpendByUser now falls
back to the default sort expression instead of interpolating undefined
into ORDER BY, findPushedEventsForSource bounds its dedup subquery by
beforeMs instead of scanning full history, and eventsForSource guards
against an invalid beforeIso producing NaN.
